Update main.h
lengthened coinbase maturity to restrict spending mining proceeds for 6 months to discourage pools, mining fluctuations based on price, and cloud mining.  Serious miners only and levels playing fields so pools are not as lucrative.  Possible that a pool could pay out immediatly and wait for the proceeds to finally post, but very unlikely.